플레이어 스탯 창 기능 구현 , 게임 시작 시 무기 선택 구현, 데이터 테이블 업데이트 (무기 종류 변경/밸런싱, 몬스터 타입 추가, 타입별 랜덤 가중치 적용/밸런싱, 업그레이드 밸런싱)

+++ FirstGame/DataTables/UpgradeTable.cpp
@@ -11,7 +11,7 @@ bool UpgradeTable::Load()
     std::vector<std::string> category  = doc.GetColumn<std::string>(2);
     std::vector<std::string> highlight = doc.GetColumn<std::string>(3);
     std::vector<std::string> title     = doc.GetColumn<std::string>(4);
-    std::vector<int>         value     = doc.GetColumn<int>(5);
+    std::vector<float>       value     = doc.GetColumn<float>(5);
     std::vector<int>         amount    = doc.GetColumn<int>(6);
     std::vector<int>         tier      = doc.GetColumn<int>(7);
